Extract clone-related items from component
